Fix deterministic hashing and caching

+++ kapitan/inputs/cache.py
@@ -1,10 +1,10 @@
 import hashlib
+import pickle
 import os
 from pathlib import Path
 from kapitan.errors import CompileError
 import logging
 from typing import Tuple
-import json
 
 logger = logging.getLogger(__name__)
 
@@ -17,6 +17,8 @@ class InputCache(object):
             self.input_cache_home = f"{cache_home}/.cache/kapitan/{input_type_name}"
         else:
             raise CompileError("Could not get cache dir: $XDG_CACHE_HOME or $HOME not set.")
+
+        self.kv_cache = {}
 
         logger.debug("Input cache home: %s", self.input_cache_home)
 
@@ -33,7 +35,7 @@ class InputCache(object):
                 try:
                     with open(cached_path, "rb") as fp:
                         logger.debug("Loading cache hit: %s", cached_path)
-                        return json.load(fp)
+                        return self.load_output(fp)
                 except FileNotFoundError:
                     pass
         return None
@@ -41,18 +43,37 @@ class InputCache(object):
     def set(self, inputs_hash, output_obj, lock_retries=2):
         cached_path, cached_path_lock, sub_path = self.hash_paths(inputs_hash)
         for retry in range(lock_retries):
+            # dont write if already exists
+            if cached_path.exists():
+                return inputs_hash
+
             if not cached_path_lock.exists():
                 sub_path.mkdir(parents=True, exist_ok=True)
                 with open(cached_path_lock, "wb") as fp:
-                    output_obj_dump = json.dumps(output_obj, sort_keys=True).encode("utf-8")
-                    logger.debug("Writing cache: %s", cached_path_lock)
-                    fp.write(output_obj_dump)
+                    logger.debug("Writing cache file: %s", cached_path_lock)
+                    self.dump_output(output_obj, fp)
                 cached_path_lock.rename(Path(str(cached_path_lock).removesuffix(".lock")))
                 logger.debug("Moved cache lock to: %s", cached_path)
 
                 return inputs_hash
         return None
 
+    def set_value(self, key, value):
+        self.kv_cache[key] = value
+
+    def get_key(self, key):
+        return self.kv_cache[key]
+
     @staticmethod
     def hash_object():
         return hashlib.blake2b(digest_size=32)
+
+    @staticmethod
+    def hash_file_digest(fp):
+        return hashlib.file_digest(fp, "blake2b")
+
+    def dump_output(self, output_obj, fp):
+        return pickle.dump(output_obj, fp)
+
+    def load_output(self, fp):
+        return pickle.load(fp)
